Add focus-within and revamp Spinbox API
Introduce focus-within support and refactor the Spinbox component.

- CSS: reflow font-family, add spinbox :focus-within outline in dark and light themes, and remove textbox outline in light theme.
- Style system: add "focus-within" pseudo-class to selector parser and StyleModifiers API (focus_within), and ensure style restyling when focus-within changes in EventContext and Context.
- Spinbox: refactor to use a numeric f64 Signal value, add min/max signals and min()/max() bindings, add on_change callback, clamp logic, keyboard keymap (arrow/home/end), disable buttons at limits, set Textbox as spinbutton role, and adjust navigability behavior.
- Datepicker and examples: update usages to the new Spinbox API and replace some custom spinbox UIs with HStack + Buttons where appropriate.

These changes enable focus-within styling and provide a more fully-featured, keyboard-friendly numeric Spinbox with min/max support and change callbacks.

+++ examples/widget_gallery/src/views/spinbox.rs
@@ -3,7 +3,7 @@ use vizia::prelude::*;
 use crate::DemoRegion;
 
 struct SpinboxData {
-    spinbox_value_1: Signal<i64>,
+    spinbox_value_1: Signal<f64>,
     // spinbox_value_2: usize,
     // spinbox_value_3_choices: Vec<SpinboxValues>,
     // spinbox_value_3: usize,
@@ -49,11 +49,11 @@ impl Model for SpinboxData {
     fn event(&mut self, _cx: &mut EventContext, event: &mut Event) {
         event.map(|e, _| match e {
             SpinboxEvent::Decrement1 => {
-                self.spinbox_value_1.update(|value| *value -= 1);
+                self.spinbox_value_1.update(|value| *value -= 1.0);
             }
 
             SpinboxEvent::Increment1 => {
-                self.spinbox_value_1.update(|value| *value += 1);
+                self.spinbox_value_1.update(|value| *value += 1.0);
             } // SpinboxEvent::Decrement2 => {
               //     if self.spinbox_value_2 != 0 {
               //         self.spinbox_value_2 -= 1;
@@ -89,7 +89,7 @@ impl Model for SpinboxData {
 }
 
 pub fn spinbox(cx: &mut Context) {
-    let spinbox_value_1 = Signal::new(99);
+    let spinbox_value_1 = Signal::new(99f64);
 
     SpinboxData {
         spinbox_value_1,
